refactor: Remove gRPC from net (#3907)
## Relevant issue(s)

Resolves #3909

## Description

This PR aims to remove the need to implement a gRPC server to handle
point to point comms. It instead makes use of the network stream
provided by LibP2P (which benefits of having all the different transport
options). This should help reduce the size of the wasm build and is the
first step in trying to remove all the Defra specific business logic
from `net`.