refactor: Remove SchemaDescription (#3842)
## Relevant issue(s)

Resolves #3698
Resolves #1520 

## Description

Removes SchemaDescription from user space.

### Schema properties are now saved in the blockstore

The global properties (previously on schema and co) are now saved in the
blockstore.

The blocks are defined in the `crdt` package in order to piggyback off
of the existing block logic, even though atm they are not true CRDTs
(they could become so later if we want). I would like to avoid a further
refactor of the existing crdt code in this PR.

Only properties that have been mutated are stored in the blocks (a
delta).

The CollectionSet concept that was largely contained within the old
`db/schema_id.go` file is now also in the blockstore. This is required
for circular subsets of collections, as they must reference each other
by their relative location within the collection set instead of the
related CollectionID, as otherwise we would be trying to form a circular
DAG (dog=>cat=>dog=>cat=>...). There is more documentation on this in
`db/collection_id.go` (renamed from the original `db/schema_id.go`).

### Schema properties are now hosted on Collection

The properties on SchemaDescription, and SchemaFieldDescription have
been moved onto CollectionVersion and CollectionFieldDescription. The
following changes were made to facilitate this:
- `IsPlaceholder` on `CollectionVersion` has been added, as we can no
longer use the lack of a `SchemaDescription` to determine whether a
migration was uploaded before the schema versions were defined.
- `IsPrimary` has been added to `CollectionFieldDefinition` as we can no
longer use the lack of a `SchemaFieldDescription` to determine whether a
field is the primary side of a relation.

This also means that `PatchSchema` has been removed, and the properties
ported over are now available via `PatchCollection`.

`PatchCollection` has gained some of the substitution logic that
`PatchSchema` had, such as referencing by collection name instead of ID.
It also activates new versions by default - given the way patching
works, I think this will make a lot more sense to users, and the old
`setActive` boolean is very confusing with edits to multiple
collections, and those that do not result in new versions. The new
version may still be explicitly set to inactive, by patching the
`IsActive` property.

The internal logic around orphan collection versions (when a migration
between versions was added before the versions existed locally) had to
change slightly because of this.

### Collection IDs are now the Block CIDs

The IDs on CollectionVersion and its children are no longer formed by
hashing the json representation of the `client` objects, instead they
are now the CIDs of the blocks. This includes `CollectionID`,
`CollectionVersion.VersionID`, `CollectionFieldDescription.FieldID`, and
`CollectionVersion.CollectionSetID`.

This gives fields a stable, deterministic, global, field ID. This
FieldID is not unique to the collection, and may be found on fields
belonging to other collections - to do otherwise would create a circular
dependency between field and collection.

### Collection Fields can now be deleted via patch

Collection Fields can now be deleted via patch - the blocking off of
this was already broken, as users were able to add a field, then toggle
the version back to the previous (essentially removing the field). This
became even more flimsy with the other changes in this PR, and fairly
easy to handle, thanks to the new `FieldID`.

Collection short field ids are indexed by FieldID now in order to
prevent their reuse, as well as just being nicer IMO.