refactor: Utilize CGO handles for C bindings (#3950)
## Relevant issue(s)

Resolves #3961
Resolves #3919
Resolves #3917
Resolves #3920

## Description

This refactor changes how the C binding package works. Before, it was
split into two packages, bridge, and logic. The bridge package is the
one that exported functions to C, while the logic package contained all
of the logic, and utilized Go primitives and structs in all of the
function signatures. This had several downsides, not least of which was
that the integration testing only ever used the Go functions. This
addresses that.

The new version uses CGO handles to be able to return `uintptr_t` that
refer to the transactions and node objects. This allows us to
consolidate everything into a single package, which now also contains
the wrapper used for integration tests. There is one trick employed to
allow this to work: the makefile build command will, at build time,
rename the package from "cbindings" to "main", do the build, and then
rename the package from "main" to "cbinding." This is to meet a
requirement of CGO: that the package being built have the name "main."

The major revision to the c bindings is also applied to identities.
Previously we passed in identity by private key, as a string. We now
utilize CGO handles to be able to pass in identity as an object. See,
also, that the return type of `IdentityNew` has been adjusted to fit
this design.

**Other fixes**
This new approach to the c binding package allows us to pass in a node
for use during integration tests in a simple, and direct way. This
results in issue 3919
(https://github.com/sourcenetwork/defradb/issues/3919#issue-3298445597)
, in which certain integration tests that were related to node identity
did not work, being solved. Those tests now behave appropriately.

Additionally, coverage test has been included, addressing issue 3917
(https://github.com/sourcenetwork/defradb/issues/3917).

Further, this PR includes an implementation for the `P2PConnect`
function, and a small adjustment to the `withRetryOnNode` function
inside `tests/integration/utils.go`. The former was the missing piece
that allows networking tests to be enabled for the C client. The ladder
addresses an oddity of the C client: the C client does not return error
objects, it returns stringified errors. So a small adjustment to some
error catching code had to be made. The outcome of these changes is that
networking tests are no longer skipped. As such, this PR addresses issue
3929 (https://github.com/sourcenetwork/defradb/issues/3920)

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Specify the platform(s) on which this was tested:
- WSL

---------

Co-authored-by: Fred Carle <fredcarle@users.noreply.github.com>