feat: Handle invalid CGO handles in C bindings (#3966)
## Relevant issue(s)

Resolves #3965 

## Description

There were many portions of the c binding code base where a pointer (CGO
handle) to a node, transaction, or identity CGO handle was passed in,
and assumed valid without any error checking. This would result in a
panic and crash if the node were actually invalid. While there is always
the chance of the user passing in an invalid handle, it needs to be
handled gracefully, which is what now happens. The functions will now
return the error as part of the `Result` object, and return the error
message to the user to help indicate what happened.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Specify the platform(s) on which this was tested:
- WSL