refactor: Standardize errors in C bindings (#3970)
## Relevant issue(s)

Resolves #3969 

## Description

The way that errors were utilized in the C bindings was somewhat
different than in most other parts of our codebase. It largely just
utilized strings with format verbs; it also used them in a lot of
locations that perhaps it should not have used them. This PR seeks to do
a few things. **One)** C binding functions should now prefer returning
the error received from the Defra function, wrapped as a `Result`
structure, where possible, instead of wrapping it as a new error type
first. **Two)** When using a new error type unique to the C binding
package, the code now avoids using strings with format verbs, instead
opting to utilize functions which return new errors. When applicable,
these functions will use `errors.NewKV` and `errors.Wrap` to attach
additional information to the error. **Three**) Errors that had been
left over, from previous commits, but which are no longer directly used
anywhere, have simply been removed.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Specify the platform(s) on which this was tested:
- WSL