chore: Make C binding return types consistent (#4002)
## Relevant issue(s)

Resolves #4001 

## Description

In all places that we returned `*C.Result` to the user, we now return
`C.Result.` This required some changes be made to the utility functions
used by the integration test wrappers, and the CGO directives at the top
of the wrapper files. With these changes, the C bindings are more
consistent, as now all returned types are by value.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [ ] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Specify the platform(s) on which this was tested:
- WSL