refactor: Make net.Host interface easier to consume (#4009)
## Relevant issue(s)

Resolves #4007

## Description

Tweaks the `net.Host` interface easier to consume by removing elements
that require direct reference to the exact concrete type.

This is required by Lens, so that a trimmed down `net.Host` interface
can be consumed without making Lens dependant on Defra. The alternative
is to extract `net` to a new repository - this may be desirable long
term, so Lens can spin up a `net.Host` instance, but even if we go that
route I'd be in favour of this change as it allows external apps to
consume the interface without a direct reference.

`net.Host.PeerInfo()` should now be the only interface member containing
a a concrete custom type, I have left it, as this function is syntax
sugar around `ID()` and `Addrs()` and is never going to be required by
anything consuming the type.

The Lens code may be viewed in this PR:
https://github.com/sourcenetwork/lens/pull/122 - see
`host-go/p2p/host.go` for the re declaration of a subset of the
`net.Host` interface.