fix: Signature selection panic (#4004)
## Relevant issue(s)

Resolves #4003 

## Description

When document/block signature are enabled, and you query for a `Commit`
type without selecting the `signature` type, it will panic.

This is caused by the `addSignatureFieldToDoc` code *always* attempting
to add the signature whenever the `Commit` type is requested, but if you
don't include the `signature` type, the mapper indexer lookup panics as
it doesn't exist.

Note: The fix is applied in two spots. Proactively to prevent adding the
signature entirely, and again reactively when trying to add the
signature.

The surprising thing is that this wasn't caught by the test suite. Turns
out, the only tests that use the `EnableSigning` either A) Never query
the `Commit` type without also including the `signature` field B) Use
the builtin `testUtils.VerifyBlockSignature` and therefore skip this bug
altogether.