fix: Passive sync with node chain (#4012)
## Relevant issue(s)

Resolves #4010 
Resolves #2744  

## Description

This PR ensures that every node in a chain of connected nodes receives a
document update if they subscribe to the appropriate topic.

The issue was caused by the nodes receiving the update message trying to
fetch the new blocks, but the nodes that they were connected to didn't
yet have it themselves. We solve this by sending the update message
again once a node completes a sync.

By fixing this, we also fixed the gql subscription missing new doc
updates when those updates came from the p2p system.

Additionally, the blockstore now as the capability to track if a block
has been merged or not. This helps to prevent unnecessarily walking the
dag upon receiving an update and trying to sync.