fix: Allow configuring node identity with embedded Defra (#4023)
## Relevant issue(s)

Resolves #4022

## Description

This PR simply exposes a node config to set the internal db node
identity config.

Our test framework couldn't catch this since it has access to the
internal package.

We had discussed making DB public again. I'll be looking into doing this
in order to prevent these kinds of problems in the future.