feat: Support Commit API on GQL Subscriptions (#4019)
## Relevant issue(s)

Resolves #4018 
Resolves #3986
Resolves #4017

## Description

Primarily adds support for the commit API on GQL subscriptions.
Additionally it fixes a bug in how we were reporting certain kind of GQL
validation errors to SSE clients. Lastly, it fixes a bug in the
`dagScanNode` planner that resulted in an infinite loop if you requested
a bad field on a `commits` query.

Additionally, not marked as a "resolved" issue as I felt it wasn't
necessary, fixed the "Dead Code" issue referenced
[here](https://github.com/sourcenetwork/defradb/pull/4004#discussion_r2351488606).

For the actual implementation, I made the conscious choice to only emit
a ***single*** event for a mutation operation (create/update) instead of
one for each individual commit block that is generated (which is N+1,
where N is the number of fields mutated). This simplifies and
streamlines the commit subscription, and you can use the `links` in the
`_version` field to access the corresponding commit blocks anyway. It
does however limit the ability to subscribe to an individual fields
updates, and instead can only subscribe to the document.

Another important note that I would like feedback on... For the
Subscription schema, I am reusing the existing `commits` type from the
Query schema (this is also how the user-defined types work as well,
single type on both the Query and Subscription). This leads to an odd
behavior in both the commits API and the collection subscription...
There are some arguments that don't do anything in *some* contexts. Most
notable, as referenced above we can't subscribe on individual fields,
but the `fieldName` argument is still there, its just silently ignored.
I would like to deal with this out of scope, as some solution might also
affect how the collection subscription stuff works.

Lastly, in the `handler_store.go:execSSESubscription()` function theres
some errors that are silently ignored, should we add some error logging
here? There wasn't any before, and my refactor there was a "plain text
refactor".

PS. Open to more suggestions for tests that can be added to the commit
subscription suite.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?
Manually, updated integration suite.

(*replace*) Describe the tests performed to verify the changes. Provide
instructions to reproduce them.

Specify the platform(s) on which this was tested:
- Ubuntu (WSL2)