feat: Add Android build support (#4034)
## Relevant issue(s)

Resolves #4033

## Description

This adds support for generating shared object files for Android, using
both arm64-v8a architecture (which is used on modern devices), and
x86_64 architecture (which is used in Android Studio emulators on
Windows.) To do this, a build script was added at
`tools/scripts/build-c-shared-android.sh` and a build command was added
inside the makefile which invokes it, `build-c-shared-android`.

To actually allow the build to work, and then run correctly, some
changes have been made to the code base. See that `lens_wasmer.go` ,
`lens_wasmtime.go` and `lens_wazero.go` now use build directives to not
build when we are targeting Android. Additionally, there is a new file,
`lens_wazero_android.go` which is similar to `lens_wazero.go` except
that it sets the DefaultLens.

Lastly, `go.mod` now has one replace directive:

```
replace github.com/wlynxg/anet => github.com/sourcenetwork/anet v0.0.0-20250417190629-7c87cba7799e
```

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Specify the platform(s) on which this was tested:
- WSL
- Windows