test: Add many-to-many integration tests (#4045)
## Relevant issue(s)

Resolves #3852

## Description

Although Defra supports the creation and usage of many-to-many
relationships by way of using join collections, there were no
integration tests that demonstrate this use. This PR addresses that by
creating two new tests:

`TestManyToMany_QueryFromJoinCollection_Succeeds` and
`TestManyToMany_QueryFromStudent_Succeeds`

These tests create and use a many-to-many relationship of students and
courses.

The first of these shows that two schemas can be connected semantically
to one another through a third schema and collection that acts as a join
table. After creating the data, the test shows that the queries can be
written bi-directionally (i.e. we can write queries that go from
student-to-courses, or that go from course-to-students.)

The second demonstrates how we can use relationship directives to allow
nested queries on objects. Specifically, it shows how we are able to
query a student to access their enrollments, and get their courses that
way (avoiding ever querying Course, or Enrollments directly.)

* Communication regarding updating our documentation on this topic is
ongoing. *

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Specify the platform(s) on which this was tested:
- WSL