refactor: Limit previous collection versions to one or zero (#4062)
## Relevant issue(s)

Resolves #4057

## Description

Limit previous collection versions to one or zero at compile time.

Previously the model supported multiple sources, but was temporarily
restricted at runtime for short-term simplification. Particularly now
that the collection is represented by a DAG in the blockstore, this is
an unwanted complication that makes a lot of things harder.