chore: Allow test framework to test message relaying (#4035)
## Relevant issue(s)

Resolves #4030 

## Description

~~This PR fixes a devex issue where bootstrap instance would not be
connected to until a message was sent over the pubsub network. This
resulted it the bootstrap instance not knowing about the new instance
until later than expected.~~

~~The workaround for this was to call `Connect` instead of (or in
addition to) defining a bootstrap instance.~~

~~With this fix, the bootstrapped instance will know about the new
instance as soon as it goes online.~~

Edit: The above was taken care of in `go-p2p` which will be brought in
from another PR. Removing the fix from this branch.

In addition to this, this PR makes it so the test tries to use a local
address instead of the loopback. This is to make sure that the DHT can
relay messages across the network. Loopback addresses are ignored by the
DHT. My assumption is that the creator of the package assumed that peers
using the loopback address would be using mDNS instead when trying to
discover other local peers.