chore: Reduce size of store prefixes (#4025)
## Relevant issue(s)

Resolves #4026 

## Description

~~This PR changes the blockstore keys to use the full cid information
(version+codec+hash instead of only hash). It also uses raw bytes
instead of a base32 encoded string which reduces storage overhead.~~

This PR no only focuses on reducing the size of the store prefixes. The
blockstore portion won't be needed shortly as Andy is moving it outside
of Defra. This PR will wait for that change to happen before getting
merged.

Using raw bytes has the drawback of being difficult to use for
debugging. A `HumanReadableKey` function was added and used with
`printStore`.