fix: Version query with delete (#4071)
## Relevant issue(s)

Resolves #4008

## Description

This PR fixed an issue where doing a version query on a document that
had been deleted cause the query to hang. This also appeared in
subscriptions since subscriptions use a version query to return the
appropriate document state.

The problem stems from the version fetcher using the in-memory store and
having a btree implementation that doesn't allow modifying the btree at
the same time as iterating over it. There is a way that this could be
done by modifying the in-memory store implementation but this PR
provides a good enough fix for the short term.