fix: Allow docs to contain nillable arrays of nillable types (#4061)
## Relevant issue(s)

Resolves #4059
Resolves #2950

## Description

For a detailed explanation and example of the problem at hand, see:
https://github.com/sourcenetwork/defradb/issues/4059

To address the issue:

In `client/value.go` there is a function called `Bytes()` which contains
pre-existing logic for properly handling arrrays of nillable types. This
logic has been duplicated into the `toMap` function in
`client/document.go`, cleanly handling the issue.

In addressing this issue, we actually find that another outstanding
issue is also addressed. We wanted to be able to execute queries that
order by aggregate fields. With the previously mentioned fix, this will
just work *on arrays*. To illustrate this, some integration tests have
been written and included with this PR.

**This is a breaking change. I document this in
`i4061-nillable-aray-docs.md`**

Because the logic of the `toMap` function has changed, existing
integration tests that explicitly utilized doc IDs, and which had
schemas containing arrays of nillable types, needed to be adjusted to
use new doc IDs. There were also tests which expected documents in a
certain default order (that is, it did not explicitly order the
results), which needed to be adjusted because of a side-effect of the
fix described above.

The following tests had to change because of hardcoded doc IDs:


`TestQueryInlineNillableFloatArray_WithDocIDMaxAndPopulatedArray_Succeeds`
`TestQueryInlineNillableFloatArray_WithDocIDAndMin_Succeeds`
`TestQueryInlineNillableFloatArray_WithDocIDAndMax_Succeeds`

`TestQueryInlineNillableFloatArray_WithDocIDMaxAndPopulatedArray_Succeeds`

The following tests had to change the expected order of its result:

`TestQueryInlineArrayWithGroupByString`
`TestQueryInlineArrayWithGroupByArray`

And the structure of the following tests' `ContainsData` block was also
reworked:

`TestSchemaAggregateSimpleCreatesUsersAverage`
`TestSchemaAggregateSimpleCreatesUsersSum`
`TestSchemaAggregateSimpleCreatesUsersCount`
`TestSchemaAggregateInlineArrayCreatesUsersSum`
`TestSchemaAggregateInlineArrayCreatesUsersAverage`

**Ordering aggregate fields of scalar values**

As mentioned, the previous work allows the aggregate functions (_sum,
_count, _min, _max, _avg) to be performed, and for the query to be
ordered in ascending or descending value. However, leaving the code like
this would result in an issue that was only half-solved. We also want
these aggregate functions, and ordering, to work on scalar values. So
additional work was done to make this work. Specifically, the
`getNumericFields` function of
`internal/request/graphql/schema/generate.go` file was modified. Then,
the `Next` function of the following functions were modified as well:
`internal/planner/average.go`, `internal/planner/max.go`,
`internal/planner/min.go`, `internal/planner/sum.go`. Each of these were
changed so as to catch the cases of scalar values, and use them. This PR
contains tests to illustrate this functionality.

**DO NOT MERGE. TESTS ARE STILL BEING WORKED ON**

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Specify the platform(s) on which this was tested:
- WSL