refactor: Update CLI Examples (#4095)
## Relevant issue(s)

Resolves #4094 

## Description

The referenced issue has more context. Basically, there are a number of
CLI examples that are embedded inside the usage fields on the cobra
command that are out of date. More importantly there was no way for us
to verify the correctness of these examples. Lastly, there was some
rendering issues on the docs website that made use of the autogenerated
CLI markdown docs.

This PR introduces a new `exampleRegistry` system and cooresponding
`EmbedCLIExample` function. This function is responsible for registering
the example with the `exampleRegistry` so that it they can be accessed
in the test suite and programatically verified for correctness. It also
adds the examples and their title to the appropriate cobra
`command.Example` field, which is more intentionally rendered, both on
the CLI `--help` message, and in the autogenerated markdown docs.

This is aligned with a general effort I have going on to utilize "Docs
as Code" methodologies when it comes to our docs and code. Especially
related to v1.

*Note*: Migrating these examples and having them run within the test
suite actually caught a bunch more examples than I already knew about
that were wrong due to previous refactors and feature changes.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?
Added unit tests to cover all the existing CLI examples.

Specify the platform(s) on which this was tested:
- Ubuntu (WSL2)