refactor: Remove `isArray` from field definition delta (#4121)
## Relevant issue(s)

Resolves #4122 

## Description

This PR simplifies the `FieldDefinitionDelta`so that it doesn't need to
specify if the field is an array type. Doing this caused the field
definition block content to change and thus affected its CID, the one of
the collection and thus the docIDs. It broke a lot of tests. Only 15
lines are production code changes. The rest is test changes.

To help in reducing some of the overhead of future similar breaking
changes, a new field was added to the `testUtils.Request` struct
(`NonOrderedResults bool`) to specify that a given result set doesn't
need to be received in the specific order defined. This should help for
results that are automatically ordered by docIDs.