fix: Glibc error with Docker container (#4129)
## Relevant issue(s)

Resolves #3953

## Description

This PR fixes the `GLIBC` error seen in with the docker image. This was
caused buy the build image using Debian 13 and the deployed image using
Debian 12.

The PR also adds a workflow so that we can release Docker images outside
of the normal release workflow. This will be useful in situations like
this where something needs to be fixed for the docker image but the rest
of the release is fine.