fix: Fix identity-related bugs in c bindings (#4145)
## Relevant issue(s)

Resolves #4091 

## Description

This pull request does a few different things, which are all related to
bugs in the way the c bindings, and the c binding integration test
wrapper, used and passed identity.

Firstly:

A new function is available, `IdentityFree`, which takes an identity
pointer, checks whether or not it is valid, and if it is valid, calls
`Delete` on the associate CGO handle. This function is exported and made
available to users, however we also now use it throughout the C client
wrapper. Previously, we were creating a CGO handle every time we
extracted an identity from a context, and made a pointer out of it, but
we were never actually deleting the handle. This is probably harmless on
the scale our tests exist, but it is a memory leak, and it is wrong.
This has been fixed.

Secondly:

Several functions never took in identity, when in reality, they were
supposed to. These are `BlockVerifySignature` ,`IndexList` , and
`IndexDrop` , and all of the P2P-related functions. Because identity was
never being passed into these functions, errors creeped in with regards
to permissions of actions. To address this, these functions now take
either a `CollectionOptions` object, or an identity pointer.

*There has been a bit of questioning around whether or not the use of
CollectionOptions is ideal in all of these locations. I think we should
have a team-wide discussion on this topic, however, I think it should be
a separate PR.*

Thirdly:

These changes have resulted in integration tests that were not behaving
correctly before, to work.

The C client type has been enabled for the following tests, and they
will pass:

`TestNAC_AdminRelation_CanIndexCreate`
`TestNAC_AdminRelation_CanIndexDrop`
`TestNAC_AdminRelation_CanIndexList`

Tests prefixed with `TestNAC_GatesP2P*`
Tests prefixed with `TestNAC_AdminRelation*`

The to-do has been removed from the following tests, as it will not (and
should not) ever pass:

`TestNAC_AdminRelation_CanVerifySignature`
`TestNAC_GatesVerifySignature_AuthorizedIdentity_AllowAccess`

And instead, the C client has been added as a supported client for the
following tests, where it will pass:

`TestNAC_AdminRelation_GoClient_CanVerifySignature`
`TestNAC_GatesVerifySignature_GoClient_AllowAccess` (renamed:
`TestNAC_GatesVerifySignature_GoClient_AuthorizedIdentity_AllowAccess`)

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Specify the platform(s) on which this was tested:
- WSL