fix: Lens-migrations-aware secondary indexes (#4114)
## Relevant issue(s)

Resolves #1597

## Description

This PR make sure the secondary indexes are up-to-date for different
collection versions and migrations.
Current implementation is not ideal as it reindexes the entire
collection on every migration or version change, but this can be
optimized only after [Backfill
feature](https://github.com/sourcenetwork/defradb/issues/296) is
implemented.

This PR also adds a new Increment lens module that can be used to
increment/decrement doc field values.