fix: Encode identity as hex in commit query (#4104)
## Relevant issue(s)

Resolves #3999 

## Description

Doing a commit query that gets the `Identity` field previously returned
a string of text that was base64-encoded. This PR adjusts it so that it
returns a hex-encoded string instead. This is so it is in a format that
is more directly-usable by developers. Note that the
VerifyBlockSignature function expects a hex-encoded string.

Changing this was a one-liner, but did result in some tests breaking,
because they expected the old form. To fix them, adjustments to the
`IdentityMatcher` have been made.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

(*replace*) Describe the tests performed to verify the changes. Provide
instructions to reproduce them.

Specify the platform(s) on which this was tested:
- WSL