test: Enable all TestSignature tests for C client (#4166)
## Relevant issue(s)

Resolves #4156 

## Description

The following tests were disabled for the C client, when they didn't
need to be. The tests work as expected. This PR simply enables them.

`TestSignature_WithEd25519KeyType_ShouldIncludeSignatureData`
`TestSignature_WithDeletedDocAndCommitQuery_ShouldIncludeSignatureData`

`TestSignature_WithUpdatedDocsAndCommitQuery_ShouldSignOnlyFirstFieldBlocks`
`TestSignature_WithCommitQuery_ShouldIncludeSignatureData`

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [ ] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

(*replace*) Describe the tests performed to verify the changes. Provide
instructions to reproduce them.

Specify the platform(s) on which this was tested:
- WSL