refactor: Pass identity consistently in C binding functions (#4180)
## Relevant issue(s)

Resolves #4179 

## Description

There were some spots across the c binding code that passed in identity
pointers as raw `uintptr_t` values. However, there were other spots that
passed in identity as a parameter in the `CollectionOptions` struct.
This PR brings consistency by making sure we always pass it in by
pointer. It removes the `identityPtr` field from the struct entirely.

This change results in various changes needing to be made to the
integration test wrapper, so this PR also takes care of that.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Specify the platform(s) on which this was tested:
- WSL